# docs/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "GrowthAgent API",
    "version": "1.0.0",
    "description": "Autonomous AI growth system with ONE deterministic gatekeeper. Built for Razorpay AI Buildathon."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3000",
      "description": "Local development & demo server"
    }
  ],
  "paths": {
...
    "/v1/admin/rules": {
      "get": {
        "summary": "Get Merchant Rules",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Current active MerchantRules and version."
          }
        }
      },
      "put": {
        "summary": "Patch Merchant Rules",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Rules updated and version bumped."
          },
          "409": {
            "description": "Version conflict or limit raise cooldown."
          }
        }
      }
    },
    "/v1/admin/rules/history": {
      "get": {
        "summary": "Get Rules Changelog History",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Immutable version history log."
          }
        }
      }
    },
...
    "/v1/demo/scenarios/{name}": {
      "post": {
        "summary": "Launch Demo Scenario",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "responses": {
          "202": {
            "description": "Scenario run launched asynchronously."
          }
        }
      }
    },
    "/v1/demo/scenarios/runs/{runId}": {
      "get": {
        "summary": "Poll Demo Scenario Run Verdict",
        "security": [
          {
            "AdminTokenAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "Self-grading assertions summary."
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "AgentKeyAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Agent-Key"
      },
      "AdminTokenAuth": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Admin-Token"
      },
      "RazorpaySignature": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Razorpay-Signature"
      }
    }
  }
}